Moving functions from init to fs in utils and updating imports
