+++ gprofiler/profilers/node.py
@@ -22,7 +22,8 @@ from websocket._core import WebSocket
 
 from gprofiler.log import get_logger_adapter
 from gprofiler.metadata.versions import get_exe_version
-from gprofiler.utils import TEMPORARY_STORAGE_PATH, add_permission_dir, resource_path
+from gprofiler.utils import TEMPORARY_STORAGE_PATH, add_permission_dir
+from gprofiler.utils.fs import resource_path
 from gprofiler.utils.process import pgrep_exe
 
 logger = get_logger_adapter(__name__)
