+++ gprofiler/profilers/perf.py
@@ -23,9 +23,10 @@ from gprofiler.metadata.application_metadata import ApplicationMetadata
 from gprofiler.profilers.node import clean_up_node_maps, generate_map_for_node_processes, get_node_processes
 from gprofiler.profilers.profiler_base import ProfilerBase
 from gprofiler.profilers.registry import ProfilerArgument, register_profiler
-from gprofiler.utils import wait_event, wait_for_file_by_prefix
+from gprofiler.utils import wait_event
 from gprofiler.utils.perf import perf_path
 from gprofiler.utils.process import is_process_basename_matching, start_process, run_process
+from gprofiler.utils.fs import wait_for_file_by_prefix
 
 logger = get_logger_adapter(__name__)
