+++ gprofiler/profilers/ruby.py
@@ -21,7 +21,7 @@ from gprofiler.log import get_logger_adapter
 from gprofiler.metadata.application_metadata import ApplicationMetadata
 from gprofiler.profilers.profiler_base import SpawningProcessProfilerBase
 from gprofiler.profilers.registry import register_profiler
-from gprofiler.utils import random_prefix, removed_path, resource_path
+from gprofiler.utils.fs import removed_path, random_prefix, resource_path
 from gprofiler.utils.process import pgrep_maps, is_process_basename_matching, process_comm, search_proc_maps, run_process
 
 logger = get_logger_adapter(__name__)
